+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * Add more information to a bunch of tooltips in the binding configuration window
 * Add spell ranks to spell autofill in Classic Era
 * Create bindings when dragging a macro into the binding configuration window
+* Add quick start import from macros
 
 ### Fixed